+++ CHANGELOG.md
@@ -20,6 +20,10 @@
   * Changes to `config.default.ini`: `SEPLOS_USE_BMS_VALUES` was replaced by `USE_BMS_DVCC_VALUES`
   * Changes to `config.default.ini`: Changed default values for Cell Voltage Current Limitation and Temperature Current Limitation
 
+* Driver version greater or equal to `v2.0.20241211dev`
+  * Changes to `config.default.ini`: `SOC_CALC_CURRENT_REPORTED_BY_BMS` was replaced by `CURRENT_REPORTED_BY_BMS`
+  * Changes to `config.default.ini`: `SOC_CALC_CURRENT_MEASURED_BY_USER` was replaced by `CURRENT_MEASURED_BY_USER`
+
 ### What's Changed
 * Added: Calculation of history values not provided by the battery by @mr-manuel
 * Added: Felicity BMS by @versager
@@ -41,7 +45,12 @@
 * Changed: The root filesystem is not mounted as read-write anymore, since overlay filesystems are used now. This allows to let the core system files untouched and to revert all changes with one command. The changes are now also persistant and do not have to be installed on every Venus OS update again by @mr-manuel
 * Changed: The setting `MODBUS_ADDRESSES` was replaced by `BATTERY_ADDRESSES` in the `config.default.ini` by @mr-manuel
 * Changed: The setting `SEPLOS_USE_BMS_VALUES` was replaced by `USE_BMS_DVCC_VALUES` in the `config.default.ini` by @mr-manuel
+* Changed: The setting `SOC_CALC_CURRENT_MEASURED_BY_USER` was replaced by `CURRENT_MEASURED_BY_USER` in the `config.default.ini` by @mr-manuel
+* Changed: The setting `SOC_CALC_CURRENT_REPORTED_BY_BMS` was replaced by `CURRENT_REPORTED_BY_BMS` in the `config.default.ini` by @mr-manuel
 * Changed: Use corrected current for both normal current and SOC calculation if `SOC_CALCULATION` is enabled by @mr-manuel
+* Changed: When `SOC_CALCULATION` is enabled, the SoC is reset to 100%, if the battery switches to float. Old options were removed by @mr-manuel
+* Removed: The setting `SOC_RESET_CURRENT` was removed in the `config.default.ini` by @mr-manuel
+* Removed: The setting `SOC_RESET_TIME` was removed in the `config.default.ini` by @mr-manuel
 
 
 ## v1.5.20241202